Fix shebang portability issue
Some versions of /usr/bin/env don't split their arguments between command name
and command arguments. This causes this script to fail on those platforms:

    /usr/bin/env: ruby -KU: No such file or directory

Ubuntu 12.04 LTS is affected by this issue.

Setting $KCODE is the equivalent of -KU flags for ruby 1.8. For ruby 1.9 this is
no longer needed and will trigger a deprecation warning.